backend/requirements.txt
fastapi>=0.111.0
mangum>=0.17.0
pydantic>=2.7.1
uvicorn>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
PyJWT>=2.8.0
requests>=2.31.0
